moved all functions to handlers
