+++ resources/chatgpt.py
@@ -1,4 +1,4 @@
-import openai, re
+import openai
 from resources import config
 from resources.database import Database
 
@@ -11,19 +11,21 @@ class ChatGPT:
         self.messages  = [
             # {"role": "system", "content": f"{self.mode}"}
         ]
-        openai.api_key = self.key
+
+        self.client = openai.OpenAI(api_key=self.key)
 
     def response_completion(self, append=True):
-        completion = openai.ChatCompletion.create(
-                model = self.model,
-                messages = self.messages,
-                temperature=0.8
-            )
+        completion = self.client.chat.completions.create(
+            model = self.model,
+            messages = self.messages,
+            temperature=0.8
+        )
         response = completion.choices[0].message.content
         if append:
             self.messages.append({"role": "assistant", "content": response})
         print('Chat:', f"\n{response}\n")
         return response
+
     
     def append_message(self, message):
         print('You:', message)
