+++ resources/translate.py
@@ -1,8 +1,8 @@
 from deep_translator import GoogleTranslator
 import traceback
 
-google = GoogleTranslator(source='auto', target='ru')   
-def get_translate(text):
+def get_translate(text, lang):
+    google = GoogleTranslator(source='auto', target=lang)   
     try:
         result = google.translate(text)
         return result
