+++ start.ps1
@@ -1,1 +1,1 @@
-Start-Process python C:\Projects\Python\aiogram-bot\main.py -WindowStyle Hidden
+Start-Process ~/AppData/Local/Programs/Python/Python311/python.exe C:\Projects\Python\aiogram-bot\main.py -WindowStyle Hidden
